requirements.txt
cat > requirements.txt << 'EOF'
streamlit
groq
chromadb
sentence-transformers
pypdf
python-dotenv
EOF